Updated generateAverageAreaSensor tests
