+++ nodes/modules/system/test/unit/domain/templates/mocks/average_illuminance_area1_sensor.yaml
@@ -1,0 +1,39 @@
+template:
+  - sensor:
+      - name: Average Illuminance Area 1
+        unit_of_measurement: lx
+        state: >-
+          {% set domains = ["sensor"] %}
+          {% set combined_states = [] %}
+          {% set all_states = states
+            | selectattr('domain', 'in', domains)
+            | selectattr('entity_id', 'in', area_entities('area1'))
+            | rejectattr('state', 'eq', 'unavailable')
+
+            | rejectattr('entity_id', 'contains', 'average')
+
+            | rejectattr('entity_id', 'contains', 'battery')
+            | list %}
+
+          {% set illuminance_states = all_states
+            | selectattr('entity_id', 'contains', 'illuminance')
+            | map(attribute='state')
+            | map('float')
+            | list %}
+          {% set combined_states = combined_states + illuminance_states %}
+
+          {% set lux_states = all_states
+            | selectattr('entity_id', 'contains', 'lux')
+            | map(attribute='state')
+            | map('float')
+            | list %}
+          {% set combined_states = combined_states + lux_states %}
+
+          {% set unique_states = combined_states | unique | list %}
+
+          {% if unique_states | length > 0 %}
+          {% set average_state = unique_states | sum / unique_states | length %}
+          {{ average_state | round(1) }}
+          {% else %}
+          Unavailable
+          {% endif %}
